+++ kinto/plugins/default_bucket/test_plugin.py
@@ -162,7 +162,8 @@ class DefaultBucketViewTest(FormattedErrorMixin, BaseWebTest,
         with mock.patch.object(self.storage, 'get',
                                wraps=self.storage.get) as patched:
             self.app.post_json('/batch', batch, headers=self.headers)
-            self.assertEqual(patched.call_count, 0)
+            # Called twice only: bucket + collection ids unicity.
+            self.assertEqual(patched.call_count, 2)
 
     def test_parent_collection_is_taken_from_the_one_checked_in_batch(self):
         # Create it first.
@@ -179,7 +180,8 @@ class DefaultBucketViewTest(FormattedErrorMixin, BaseWebTest,
         with mock.patch.object(self.storage, 'get',
                                wraps=self.storage.get) as patched:
             self.app.post_json('/batch', batch, headers=self.headers)
-            self.assertEqual(patched.call_count, 0)
+            # Called twice only: bucket + collection ids unicity.
+            self.assertEqual(patched.call_count, 2)
 
     def test_collection_id_is_validated(self):
         collection_url = '/buckets/default/collections/__files__/records'