+++ CHANGELOG.rst
@@ -18,7 +18,8 @@ This document describes changes between each past release.
 - Redis listener is not part of the core anymore. (#712)
   Use ``kinto.event_listeners.redis.use = kinto_redis.listeners`` instead of
   ``kinto.event_listeners.redis.use = kinto.core.listeners.redis``
-  
+- Notion of unique fields was dropped from ``kinto.core`` resources.
+
 **Protocol**
 
 - Added a ``/__version__`` endpoint with the version that has been deployed. (#747)